// client/src/enums/interestItem.ts
export enum InterestItemText {
    Music = "Music",
    Films = "Films",
    IT = "IT",
    Reading = "Reading",
    Art = "Art",
    Gaming = "Gaming",
    Sports = "Sports",
    Fashion = "Fashion",
    Science = "Science",
    Nature = "Nature",
    History = "History",
    Economics = "Economics",
    Marketing = "Marketing",
    Beauty = "Beauty",
    Animals = "Animals",
    SocialMedia = "Social Media",
    Cultures = "Cultures",
    Travelling = "Travelling",
    Medicine = "Medicine",
    Cooking = "Cooking",
    Fitness = "Fitness",
    ContentMaking = "Content Making",
    Cartoons = "Cartoons",
}